fix: replace images
