+++ components/icons/website.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 function Website() {
   return (
     <svg xmlns='http://www.w3.org/2000/svg' width='21' height='21' viewBox='0 0 21 21' fill='none'>
-      <g opacity='0.5'>
+      <g opacity='1'>
         <path
           d='M10.4992 2.2002C15.0834 2.2002 18.7992 5.91604 18.7992 10.5002C18.7992 15.0843 15.0834 18.8002 10.4992 18.8002C5.91507 18.8002 2.19922 15.0843 2.19922 10.5002C2.19922 5.91604 5.91507 2.2002 10.4992 2.2002ZM9.86076 11.1387L7.32053 11.1391C7.33841 11.5477 7.37289 11.9452 7.42269 12.3296L7.46227 12.6148L7.51164 12.9191C7.90579 15.1716 8.82304 16.8644 9.86118 17.3683L9.86076 11.1387ZM13.6779 11.1391L11.1377 11.1387V17.3683C12.1524 16.875 13.0518 15.2461 13.4596 13.0685L13.4868 12.9191L13.5362 12.6152C13.6095 12.126 13.6568 11.6333 13.6779 11.1391ZM6.04233 11.1391H3.50466C3.73706 13.7142 5.36003 15.8892 7.61678 16.9065C7.17539 16.2225 6.81146 15.3819 6.54416 14.4318L6.48372 14.2105L6.41349 13.9279L6.34879 13.6393C6.17603 12.8161 6.07351 11.9797 6.04233 11.1391ZM17.4938 11.1391H14.9557C14.9275 11.8901 14.843 12.6379 14.7028 13.3763L14.6492 13.6393L14.5845 13.9279L14.5143 14.2109C14.2419 15.2512 13.8567 16.1702 13.3812 16.9065C15.638 15.8892 17.261 13.7142 17.4929 11.1391H17.4938ZM7.61678 4.09345L7.55974 4.11941C5.33194 5.14776 3.73493 7.30789 3.50466 9.86173H6.04233C6.07298 9.08281 6.15938 8.33155 6.29516 7.62414L6.34879 7.36109L6.41349 7.07251L6.48372 6.78946C6.75613 5.74919 7.14134 4.83023 7.61678 4.09387V4.09345ZM9.86076 3.63205C8.84901 4.12324 7.95176 5.74493 7.54229 7.91357L7.51164 8.08128L7.46227 8.38518C7.38894 8.87451 7.34163 9.36739 7.32053 9.86173H9.86076V3.63205ZM13.3817 4.09387L13.4327 4.17474C13.8499 4.84172 14.1955 5.65087 14.4521 6.56004L14.5147 6.78988L14.5849 7.07251L14.6496 7.36109C14.8165 8.14427 14.9216 8.98534 14.9561 9.86173H17.4938C17.2618 7.28618 15.6388 5.11115 13.3817 4.0943V4.09387ZM11.1377 3.63205V9.86173H13.6779C13.6602 9.45219 13.6247 9.04361 13.5715 8.63716L13.5362 8.38518L13.4868 8.08128C13.0926 5.82921 12.1758 4.13644 11.1381 3.63205H11.1377Z'
           fill='currentColor'
