+++ components/product/use-cases-icons.tsx
@@ -86,27 +86,47 @@ export function Sliders() {
   );
 }
 
-export function ChatTexts() {
+export function CustomViewsIcon() {
   return (
     <svg xmlns='http://www.w3.org/2000/svg' width='21' height='20' viewBox='0 0 21 20' fill='none'>
-      <path
-        fillRule='evenodd'
-        clipRule='evenodd'
-        d='M2.74114 4.11612C2.97556 3.8817 3.2935 3.75 3.62502 3.75H17.375C17.7065 3.75 18.0245 3.8817 18.2589 4.11612C18.4933 4.35054 18.625 4.66848 18.625 5V15C18.625 15.3315 18.4933 15.6495 18.2589 15.8839C18.0245 16.1183 17.7065 16.25 17.375 16.25H6.98944L6.90829 16.2785L4.42192 18.3643C4.23982 18.515 4.0187 18.611 3.78425 18.6411C3.5498 18.6712 3.31162 18.6342 3.09734 18.5345C2.88306 18.4347 2.70146 18.2762 2.57361 18.0774C2.44577 17.8786 2.37691 17.6476 2.37504 17.4112L2.375 17.4062L2.37502 5C2.37502 4.66848 2.50672 4.35054 2.74114 4.11612ZM17.375 5L3.62502 5V17.4013L6.19208 15.2477C6.24972 15.1994 6.31561 15.1618 6.38659 15.1369L6.67565 15.0353C6.74222 15.0119 6.81227 15 6.88283 15H17.375V5Z'
-        fill='currentColor'
-      />
-      <path
-        fillRule='evenodd'
-        clipRule='evenodd'
-        d='M7.375 8.75C7.375 8.40482 7.65482 8.125 8 8.125H13C13.3452 8.125 13.625 8.40482 13.625 8.75C13.625 9.09518 13.3452 9.375 13 9.375H8C7.65482 9.375 7.375 9.09518 7.375 8.75Z'
-        fill='currentColor'
-      />
-      <path
-        fillRule='evenodd'
-        clipRule='evenodd'
-        d='M7.375 11.25C7.375 10.9048 7.65482 10.625 8 10.625H13C13.3452 10.625 13.625 10.9048 13.625 11.25C13.625 11.5952 13.3452 11.875 13 11.875H8C7.65482 11.875 7.375 11.5952 7.375 11.25Z'
-        fill='currentColor'
-      />
+      <g opacity='1'>
+        <path
+          d='M12.0049 4.51718H17.8424C18.0969 4.51718 18.3032 4.31086 18.3032 4.05638C18.3032 3.8019 18.0969 3.59558 17.8424 3.59558H12.0049C11.7504 3.59558 11.5441 3.8019 11.5441 4.05638C11.5441 4.31084 11.7504 4.51718 12.0049 4.51718Z'
+          fill='currentColor'
+          stroke='currentColor'
+          stroke-width='0.205332'
+        />
+        <path
+          d='M17.8424 6.31738H12.0049C11.7504 6.31738 11.5441 6.5237 11.5441 6.77818C11.5441 7.03266 11.7504 7.23898 12.0049 7.23898H17.8424C18.0969 7.23898 18.3032 7.03266 18.3032 6.77818C18.3032 6.5237 18.0969 6.31738 17.8424 6.31738Z'
+          fill='currentColor'
+          stroke='currentColor'
+          stroke-width='0.205332'
+        />
+        <path
+          d='M7.99349 1.69738H4.77028C3.62571 1.69738 2.69787 2.62523 2.69787 3.76979V6.993C2.69787 8.13757 3.62571 9.06541 4.77028 9.06541H7.99349C9.13805 9.06541 10.0659 8.13757 10.0659 6.993V3.76979C10.0659 2.62524 9.13804 1.69738 7.99349 1.69738ZM9.14429 6.993C9.14429 7.62857 8.62905 8.14381 7.99349 8.14381H4.77028C4.13472 8.14381 3.61948 7.62857 3.61948 6.993V3.76979C3.61948 3.13423 4.13472 2.61899 4.77028 2.61899H7.99349C8.62905 2.61899 9.14429 3.13423 9.14429 3.76979V6.993Z'
+          fill='currentColor'
+          stroke='currentColor'
+          stroke-width='0.205332'
+        />
+        <path
+          d='M17.8424 12.8354H12.0049C11.7504 12.8354 11.5441 13.0418 11.5441 13.2962C11.5441 13.5507 11.7504 13.757 12.0049 13.757H17.8424C18.0969 13.757 18.3032 13.5507 18.3032 13.2962C18.3032 13.0418 18.0969 12.8354 17.8424 12.8354Z'
+          fill='currentColor'
+          stroke='currentColor'
+          stroke-width='0.205332'
+        />
+        <path
+          d='M17.8424 15.5572H12.0049C11.7504 15.5572 11.5441 15.7636 11.5441 16.018C11.5441 16.2725 11.7504 16.4789 12.0049 16.4789H17.8424C18.0969 16.4789 18.3032 16.2725 18.3032 16.018C18.3032 15.7636 18.0969 15.5572 17.8424 15.5572Z'
+          fill='currentColor'
+          stroke='currentColor'
+          stroke-width='0.205332'
+        />
+        <path
+          d='M7.99349 10.9373H4.77028C3.62571 10.9373 2.69787 11.8651 2.69787 13.0097V16.2329C2.69787 17.3774 3.62571 18.3053 4.77028 18.3053H7.99349C9.13805 18.3053 10.0659 17.3774 10.0659 16.2329V13.0097C10.0659 11.8651 9.13804 10.9373 7.99349 10.9373ZM9.14429 16.2329C9.14429 16.8684 8.62905 17.3837 7.99349 17.3837H4.77028C4.13472 17.3837 3.61948 16.8684 3.61948 16.2329V13.0097C3.61948 12.3741 4.13472 11.8589 4.77028 11.8589H7.99349C8.62905 11.8589 9.14429 12.3741 9.14429 13.0097V16.2329Z'
+          fill='currentColor'
+          stroke='currentColor'
+          stroke-width='0.205332'
+        />
+      </g>
     </svg>
   );
 }
